board: ti: j721e: evm: Split code to isolate eeprom detected logic
Move code out of evm.c that is based on evm specific checks for board
variants and initialization sequences there of.
